Rag-gemini/requirements.txt
streamlit
google-generativeai 
python-dotenv 
langchain
PyPDF2 
chromadb 
faiss-cpu 
langchain-google-genai
langchain_community